split onnx conversion logic to multiple files, clean up code
