+++ src/pquant/core/keras/layers.py
@@ -1957,8 +1957,6 @@ class PQMultiheadAttention(keras.layers.Layer):
 
     @staticmethod
     def _split_qkv_shapes(input_shape):
-        # Resolve (query, key, value) shapes from either a single shape (self-attention)
-        # or a list/tuple of per-input shapes; missing key/value fall back to query/key.
         if isinstance(input_shape, (list, tuple)) and len(input_shape) > 0 and isinstance(input_shape[0], (list, tuple)):
             q_shape = input_shape[0]
             k_shape = input_shape[1] if len(input_shape) > 1 else q_shape
@@ -1968,20 +1966,12 @@ class PQMultiheadAttention(keras.layers.Layer):
         return q_shape, k_shape, v_shape
 
     def compute_output_shape(self, input_shape):
-        # Provide static output shapes so Keras does not run call() symbolically for
-        # shape inference (which would build the quantized/pruned sublayers inside a
-        # scratch graph and fail).  Mirrors the (output, avg_attn_weights) tuple that
-        # call() returns: output is (B, Tq, embed_dim), attn weights are (B, Tq, Tk).
         q_shape, k_shape, _ = self._split_qkv_shapes(input_shape)
         batch, tgt_len = q_shape[0], q_shape[1]
         src_len = k_shape[1]
         return (batch, tgt_len, self.embed_dim), (batch, tgt_len, src_len)
 
     def build(self, input_shape):
-        # Build the projection/softmax sublayers explicitly.  Without this Keras would
-        # try to auto-build the layer by tracing call() in a scratch FuncGraph, which
-        # fails for the quantized/pruned PQDense sublayers (their build() creates tensors
-        # that escape the scratch graph).  Mirrors how PQDense itself defines build().
         q_shape, k_shape, v_shape = self._split_qkv_shapes(input_shape)
         self.q_proj.build(q_shape)
         self.k_proj.build(k_shape)
