+++ src/pquant/core/keras/onnx/__init__.py
@@ -1,0 +1,5 @@
+from pquant.core.keras.onnx.convert_to_onnx import (
+    convert_to_onnx,
+)
+
+__all__ = ["convert_to_onnx"]
